refactor(tools): split output parsing from formatting
